ISSUE-39 ISSUE-181 #time 3h
- Implementação de um novo algoritmo para a USEI07
- Ajustes em algumas funcionalidades

+++ src/main/java/isep/lapr3/g094/application/controller/GraphController.java
@@ -95,7 +95,7 @@ public class GraphController {
         return service.getLatestFileFromDirectory(string);
     }
 
-    public void maximizedPath(String idOrigem, LocalTime time, double autonomy, int velocity, Boolean bigGraph) {
+    public void maximizedPath(String idOrigem, LocalTime time, int autonomy, int velocity, Boolean bigGraph) {
         service.maximizedPath(idOrigem, time, autonomy, velocity, bigGraph);
     }
